requirements.txt
seaborn
pandas
numpy
ipywidgets
scikit-learn
streamlit_extras
streamlit_option_menu
plotly
